backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
pydantic==2.6.1
python-multipart==0.0.9
chromadb==0.4.22
langchain-community==0.0.27
pypdf==4.0.1
markdown==3.5.2
python-magic==0.4.27 